Made Statistics a class
